+++ transaction-service-api/src/main/java/pl/adamboguszewski/transaction/service/service/api/transaction/CreateTransactionResponse.java
@@ -1,0 +1,4 @@
+package pl.adamboguszewski.transaction.service.service.api.transaction;
+
+public interface CreateTransactionResponse {
+}
