+++ transaction-service-api/src/main/java/pl/adamboguszewski/transaction/service/service/api/transaction/CreateTransactionSuccessResponse.java
@@ -1,4 +1,4 @@
-package pl.adamboguszewski.transactionservice.service.api.transaction;
+package pl.adamboguszewski.transaction.service.service.api.transaction;
 
 import lombok.Value;
