--- transaction-service-api/src/main/java/pl/adamboguszewski/transactionservice/service/api/transaction/CreateTransactionResponse.java
@@ -1,4 +1,0 @@
-package pl.adamboguszewski.transactionservice.service.api.transaction;
-
-public interface CreateTransactionResponse {
-}
