+++ transaction-service/src/main/java/pl/adamboguszewski/transaction/service/TransactionServiceApplication.java
@@ -1,4 +1,4 @@
-package pl.adamboguszewski.transactionservice;
+package pl.adamboguszewski.transaction.service;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
