+++ transaction-service/src/main/java/pl/adamboguszewski/transaction/service/application/ApplicationController.java
@@ -1,9 +1,9 @@
-package pl.adamboguszewski.transactionservice.application;
+package pl.adamboguszewski.transaction.service.application;
 
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-import pl.adamboguszewski.transactionservice.service.api.transaction.CreateTransactionRequest;
+import pl.adamboguszewski.transaction.service.service.api.transaction.CreateTransactionRequest;
 
 @RestController
 @RequestMapping
