+++ transaction-service/src/main/java/pl/adamboguszewski/transaction/service/application/Transaction.java
@@ -1,7 +1,7 @@
-package pl.adamboguszewski.transactionservice.application;
+package pl.adamboguszewski.transaction.service.application;
 
 import lombok.Value;
-import pl.adamboguszewski.transactionservice.service.api.transaction.CreateTransactionRequest;
+import pl.adamboguszewski.transaction.service.service.api.transaction.CreateTransactionRequest;
 
 import java.util.ArrayList;
 import java.util.List;
