+++ transaction-service/src/main/java/pl/adamboguszewski/transaction/service/application/TransactionProduct.java
@@ -1,9 +1,9 @@
-package pl.adamboguszewski.transactionservice.application;
+package pl.adamboguszewski.transaction.service.application;
 
 import java.util.UUID;
 
 import lombok.Value;
-import pl.adamboguszewski.transactionservice.service.api.transaction.CreateTransactionRequest;
+import pl.adamboguszewski.transaction.service.service.api.transaction.CreateTransactionRequest;
 
 @Value
 public class TransactionProduct {
