+++ transaction-service/src/test/java/pl/adamboguszewski/transaction/service/test/TransactionServiceTransactionServiceApplicationTests.java
@@ -1,4 +1,4 @@
-package pl.adamboguszewski.transactionservice.test;
+package pl.adamboguszewski.transaction.service.test;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
